@@ -1,4 +1,4 @@
 pub mod base;
-pub mod entities;
+pub mod models;
 pub mod repositories;
 pub mod services;
