@@ -1,9 +1,12 @@
+use crate::settings::docs_setting::ApiDoc;
 use actix_web::{get, App, HttpServer, Responder};
 use dotenv::dotenv;
+use utoipa::OpenApi;
+use utoipa_swagger_ui::SwaggerUi;
 
 mod common;
-mod config;
 mod modules;
+mod settings;
 
 #[get("/")]
 async fn index() -> impl Responder {
@@ -13,12 +16,15 @@ async fn index() -> impl Responder {
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
     dotenv().ok();
-    config::mongodb::setup_mongo().await.unwrap();
+    settings::database_setting::setting().await.unwrap();
 
     HttpServer::new(|| {
         App::new()
             .service(index)
             .configure(|cfg| modules::todos::configure(cfg))
+            .service(
+                SwaggerUi::new("/docs/{_:.*}").url("/api-docs/openapi.json", ApiDoc::openapi()),
+            )
     })
     .bind("0.0.0.0:3000")?
     .run()
