@@ -1,10 +1,10 @@
 use actix_web::web;
 
-mod controllers;
-mod dto;
-mod entities;
-mod repositories;
-mod services;
+pub mod controllers;
+pub mod dto;
+pub mod models;
+pub mod repositories;
+pub mod services;
 
 pub fn configure(config: &mut web::ServiceConfig) {
     controllers::todo_controller::configure(config)
